+++ src/main/java/kr/co/d_erp/repository/oracle/ItemCatRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import kr.co.d_erp.dtos.CatDto;
 
-public interface ItemCatRepository extends JpaRepository<CatDto, Integer> {
-	List<CatDto> findByParentIdxOrderByCatIdx(Integer parentIdx);
+public interface ItemCatRepository extends JpaRepository<CatDto, Long> {
+	List<CatDto> findByParentIdxOrderByCatIdx(Long parentIdx);
     List<CatDto> findByParentIdxIsNullOrderByCatIdx();
 }
