+++ src/main/java/kr/co/d_erp/repository/oracle/ItemUnitRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import kr.co.d_erp.dtos.UnitForItemDto;
 
-public interface ItemUnitRepository extends JpaRepository<UnitForItemDto, Integer> {
+
+public interface ItemUnitRepository extends JpaRepository<UnitForItemDto, Long> {
 }
